requirements.txt
alembic==1.7.5
anyio==3.4.0
asgiref==3.4.1
asyncclick==8.0.1.3
asyncpg==0.25.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
databases==0.5.3
env-config==1.10.0
fastapi==0.70.1
greenlet==1.1.2
h11==0.12.0
idna==3.3
importlib-metadata==4.10.0
importlib-resources==5.4.0
Mako==1.1.6
MarkupSafe==2.0.1
passlib==1.7.4
password==0.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
python-decouple==3.4
python-dotenv==0.19.2
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.29
starlette==0.16.0
typing_extensions==4.0.1
uvicorn==0.16.0
watchgod==0.7
websockets==10.1
zipp==3.7.0
